Add multi-region support
This patch adds a new class, 'regions' to the
openstack folder, and two defined types that create
keystone endpoints for an entire region (region_auth)
and for a single service (service_auth).

This class is opinionated, and does not attempt to make
every configuration option available, instead preferring
to create sensibly named resources on the default ports
for all services.