Add support for dashboard SSL
This patch adds a simple option to enable SSL on the dashboard
that will have port 443 at the front and port 80 at the back,
plus a redirect for port 80 to port 443.

The SSL cert path should be specified either as part of a list
of bind_options or via the convenience parameter provided.

An SSL port has also been opened in the dashboard firewall rule.